Various changes related to MACE convolution. (#40)
* Patched convolution operators.

* Added some more bookkeeping to the MACE driver.

* Generated trace, ready to modify the atomic_data structure in mace_dev to use deterministic algorithm.

* NVCC no longer required to build our library.

* Removed hardcoded path.

* Minor modifications to CMakeLists.txt, removed env.sh.

* Checkpoint, need to fix bug in loop unroller.

* Preparing to re-benchmark node-edge tensor products.

* Checkpoint, ready to re-benchmark.

* Handled benzene case without compilation for e3nn.

* Removed torch overhead for roofline profiling of cuE.

* Fixed correctness.

* Switched to cuBIN API.

* Works in custom conda environment.

* Verified that benchmark_conv is in a working state.

* Verified UVW benchmarks working.

* Changed class name back.

+++ src/benchmark/tpp_creation_utils.py
@@ -1,3 +1,5 @@
+import numpy as np
+
 from typing import Iterator, Optional
 from src.implementations.e3nn_lite import Irrep, Irreps, TPProblem
 
@@ -122,7 +124,9 @@ class ChannelwiseTPP(TPProblem):
         irreps_in1: Irreps,
         irreps_in2: Irreps,
         irreps_out: Irreps,
-        label: Optional[str] = None):
+        label: Optional[str] = None,
+        irrep_dtype=np.float32,
+        weight_dtype=np.float32):
 
         trainable = True
         irreps1 = Irreps(irreps_in1)
@@ -152,7 +156,9 @@ class ChannelwiseTPP(TPProblem):
         super().__init__(irreps1, irreps2, irreps_out, instructions,
             internal_weights=False,
             shared_weights=False,
-            label=label)
+            label=label,
+            irrep_dtype=irrep_dtype,
+            weight_dtype=weight_dtype)
 
 class SingleInstruction(TPProblem):
     def __init__(
